# swagger.yaml
openapi: 3.0.0
info:
  title: SMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /test/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Test
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Test
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    post:
      summary: CRUD ACTION update
      tags:
        - Test
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: lessonno
          in: path
          description: Lesson Number
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Test"
  /test/findBy_Class/{key}:
    get:
      summary: CRUD ACTION findBy_Class
      tags:
        - Test
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
  /test/findBy_Lessons/{key}:
    get:
      summary: CRUD ACTION findBy_Lessons
      tags:
        - Test
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
  /test/findBy_Subjects/{key}:
    get:
      summary: CRUD ACTION findBy_Subjects
      tags:
        - Test
      parameters:
        - name: name
          in: path
          description: Subject Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
  /test/findBy_Teachers/{key}:
    get:
      summary: CRUD ACTION findBy_Teachers
      tags:
        - Test
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
...
  /test/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Test
      parameters:
        - name: class
          in: path
          description: Class Number
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /test/{id}/get_Lessons:
    get:
      summary: CRUD ACTION get_Lessons
      tags:
        - Test
      parameters:
        - name: lessonno
          in: path
          description: Lesson Number
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /test/{id}/get_Subjects:
    get:
      summary: CRUD ACTION get_Subjects
      tags:
        - Test
      parameters:
        - name: name
          in: path
          description: Subject Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /test/{id}/get_Teachers:
    get:
      summary: CRUD ACTION get_Teachers
      tags:
        - Test
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Test:
      properties:
        _id:
          type: string
        averageresult:
          type: number
        class:
          type: number
        date:
          type: integer
        lessonno:
          type: number
        maxmarks:
          type: number
        period:
          type: number
        subjectname:
          type: string
        teacher:
          type: string
      required:
        -  _id
        - class
        - date
        - lessonno
        - subjectname
        - teacher
...